requirements.txt
Django==1.9.7
django-bootstrap3==7.0.1
django-ordered-model==1.1.0
easy-thumbnails==2.3
mailer==0.8.1
Pillow==3.3.0
python-magic==0.4.12
